move body freeze logic to useModal
