+++ app/layouts/base.vue
@@ -4,6 +4,7 @@
 </template>
 
 <script lang="ts" setup>
+useModal()
 useViewportUnitSizes(true)
 
 if (process.env.NODE_ENV === 'development') {
